[INFRA]: Core 모듈 폴더링
