+++ GaeManda/Projects/Core/GMDExtensions/Rx/Single+Extension.swift
