+++ GaeManda/Projects/Core/GMDExtensions/UIKit/UIDevice+Extension.swift
