+++ GaeManda/Projects/Presentation/GMDProfile/Implementations/Widgets/DogProfileScrollView.swift
@@ -248,12 +248,12 @@ private extension DogProfileScrollView {
 			.bind(to: weightTextField.rx.text)
 			.disposed(by: disposeBag)
 		
-		// suffix부분에 커서가 이동이 안되도록 해줌.
-		weightTextField.textField.rx.cursorChanged
-			.bind(with: self) { owner, _ in
-				owner.weightTextField.textField.moveCusorLeftTo(suffix: "kg")
-			}
-			.disposed(by: disposeBag)
+//		 suffix부분에 커서가 이동이 안되도록 해줌.
+//		weightTextField.textField.rx.cursorChanged
+//			.bind(with: self) { owner, _ in
+//				owner.weightTextField.textField.moveCusorLeftTo(suffix: "kg")
+//			}
+//			.disposed(by: disposeBag)
 	}
 		
 	func textViewBind() {
